# [APO-1843] Optimize pydantic dict serialization by skipping annotation metadata conversion