--- a/src/vellum/client/core/pydantic_utilities.py
+++ b/src/vellum/client/core/pydantic_utilities.py
@@ -158,7 +158,14 @@ class UniversalBaseModel(pydantic.BaseModel):
 
             dict_dump = super().dict(**kwargs_with_defaults_exclude_unset_include_fields)
 
-        return convert_and_respect_annotation_metadata(object_=dict_dump, annotation=self.__class__, direction="write")
+        # return convert_and_respect_annotation_metadata(object_=dict_dump, annotation=self.__class__, direction="write")
+        # TODO: also need to update postprocessing.ts in vellum-client-generator look for prior art in this file there
+        # SDK preview should show no diffs
+        if self.__class__.__name__ in annotated_types:
+            return convert_and_respect_annotation_metadata(
+                object_=dict_dump, annotation=self.__class__, direction="write"
+            )
+        return dict_dump
 
 
 def _union_list_of_pydantic_dicts(source: List[Any], destination: List[Any]) -> List[Any]: