The purpose of this PR is to dramatically improve workflow serialization performance by conditionally skipping expensive annotation metadata conversion for types that don't require it.

## Performance Impact

Profiling data shows significant improvements across different workflow scenarios:

| Scenario | Before | After | Improvement |
|----------|--------|-------|-------------|
| Synthetic MapNode (10K items) | 1.573s | 0.014s | **99% faster** (−1.559s) |
| serialize_bug_repro workflow | 0.080s | 0.063s | 21% faster (−0.017s) |
| Simple workflow baseline | 0.013s | 0.011s | 15% faster (−0.002s) |

## Changes Made

- Modified `UniversalBaseModel.dict()` in `src/vellum/client/core/pydantic_utilities.py` to conditionally skip `convert_and_respect_annotation_metadata` for types not in `annotated_types`
- The optimization specifically targets large nested structures where repeated `typing.get_origin` checks were dominating execution time
- No changes to behavior for types that actually require annotation metadata conversion

## Technical Details

**Before**: `convert_and_respect_annotation_metadata` was called unconditionally on every dict serialization, performing expensive type introspection even when unnecessary.

**After**: Added a guard to only perform annotation metadata conversion for types that are in the `annotated_types` set, avoiding the overhead for standard types.

The most dramatic improvement is seen in the synthetic MapNode workload where the expensive Pydantic recursion on 10K list elements is eliminated, cutting runtime by ~99%.

## Testing

- Existing test suite passes
- Profiling confirms no regression on simple workflows
- Large nested structure serialization shows dramatic improvement

---
*This PR was created with Claude Code*